lenovo-accessory: Remove unnecessary guint8 cast on chunk data size
The dfu_file function takes gsize datasz, and all other callers pass
fu_chunk_get_data_sz() without narrowing. The cast silently truncates
if the value ever exceeds 255.

+++ plugins/lenovo-accessory/fu-lenovo-accessory-hid-bootloader.c
@@ -46,7 +46,7 @@ fu_lenovo_accessory_hid_bootloader_write_files(FuLenovoAccessoryHidBootloader *s
 						       file_type,
 						       fu_chunk_get_address(chk),
 						       fu_chunk_get_data(chk),
-						       (guint8)fu_chunk_get_data_sz(chk),
+						       fu_chunk_get_data_sz(chk),
 						       error))
 			return FALSE;
 		fu_progress_step_done(progress);
